add PollCount option to poll_items
This change adds a PollCount option as a 6th parameter in the poll_items list. PollCount can be used to only refresh items on every n-th poll, e.g. if PollCount is not specified or set to 1, the behavior doesn't change to the current implementation. If poll count is set to 2, this item will only get polled every 2nd time. So if you set the poll interval to 10 seconds and the poll count to 6, it will be polled every 60 seconds approximately (not accounting for wait times of the script).

This change helps to reduce the load on the Optolink interface, by being able to configure only polling certain (non-frequently changed) items less often. E.g. polling the "jahresarbeitszahl" every 60 seconds doesn't make sense, so polling it once / twice a day is enough.

After the script is restarted, every item will be polled once.